+++ faststack/check_scipy.py
@@ -1,6 +1,0 @@
-try:
-    import scipy.ndimage
-
-    print("scipy available")
-except ImportError:
-    print("scipy NOT available")